[✨ feat] 상단 Navbar 추가
1. Vue Router 설치 및 적용
2. home, todos 페이지로 라우팅 적용
