+++ src/pages/index.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>Home Page</div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
